LAMPPrimerDesign/requirements.txt
biopython==1.78
primer3-py==0.6.0
pandas==1.1.3
numpy==1.19.1
scikit-learn==0.23.2
tqdm==4.50.2
matplotlib==3.3.1
seaborn==0.11.0
